requirements.txt
azure-functions==1.18.0
azure-identity
azure-storage-blob
httpx
aiohttp
pydantic
pydantic-settings
opencensus-ext-azure
pytz
seqlog>=0.3.28
redis>=5.0.0
